requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-core
langchain-groq
langchain-huggingface
huggingface-hub
python-dotenv
requests
